requirements.txt
pillow
numpy<2
opencv-contrib-python
tensorflow
tf-keras
torch
torchvision
safetensors
ultralytics
matplotlib
onnxruntime
mediapipe
kornia
deepface
setuptools
